upload GenieAI

+++ vkge/GinieAI.py
@@ -157,7 +157,7 @@ class GinieAI:
                          for i in range(noofclusters)]
             ##These nodes will assign the centroid Variables the appropriate
             ##values
-            centroid_value = tf.placeholder("float64", [dim])
+            centroid_value = tf.placeholder("float32", [dim],dtype=tf.float32)
             cent_assigns = []
             for centroid in centroids:
                 cent_assigns.append(tf.assign(centroid, centroid_value))
@@ -167,7 +167,7 @@ class GinieAI:
             assignments = [tf.Variable(0) for i in range(len(vectors))]
             ##These nodes will assign an assignment Variable the appropriate
             ##value
-            assignment_value = tf.placeholder("int32")
+            assignment_value = tf.placeholder("int32",dtype=tf.int32)
             cluster_assigns = []
             for assignment in assignments:
                 cluster_assigns.append(tf.assign(assignment,
@@ -175,22 +175,22 @@ class GinieAI:
 
             ##Now lets construct the node that will compute the mean
             # The placeholder for the input
-            mean_input = tf.placeholder("float", [None, dim])
+            mean_input = tf.placeholder("float", [None, dim],dtype=tf.float32)
             # The Node/op takes the input and computes a mean along the 0th
             # dimension, i.e. the list of input vectors
             mean_op = tf.reduce_mean(mean_input, 0)
 
             ##Node for computing Euclidean distances
             # Placeholders for input
-            v1 = tf.placeholder("float", [dim])
-            v2 = tf.placeholder("float", [dim])
+            v1 = tf.placeholder("float", [dim],dtype=tf.float32)
+            v2 = tf.placeholder("float", [dim],dtype=tf.float32)
             euclid_dist = tf.sqrt(tf.reduce_sum(tf.pow(tf.sub(
                 v1, v2), 2)))
 
             ##This node will figure out which cluster to assign a vector to,
             ##based on Euclidean distances of the vector from the centroids.
             # Placeholder for input
-            centroid_distances = tf.placeholder("float", [noofclusters])
+            centroid_distances = tf.placeholder("float", [noofclusters],dtype=tf.float32)
             cluster_assignment = tf.argmin(centroid_distances, 0)
 
             ##INITIALIZING STATE VARIABLES
